Added flag to indicate if polygon is created from sensor measurements
This allows to differentiate polygons created from sensor measurements
and polygons created from a union. Polygons created from sensor
measurements have a vector maximum ranges which indicates for every
measurement point if it was observer at the maximum sensor range or not.
For a polygon created with a union operation, this information is not
available. To determine frontier/obstacles/free space, these information
can be used.